+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Added
 
 - Add `no-charge` flag to create comment/email commands
+- Add comment and label filters to `get_statistics`
+- Add timeseries to `get_statistics`
 
 
 ## Added